requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
pypdf
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
huggingface-hub==0.14.1
sentence-transformers==2.2.2
torch
